Refactor calendar grid
